backend/requirements.txt
aiofiles==24.1.0
aiosmtplib==5.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==3.2.0
blinker==1.9.0
brotli==1.2.0
certifi==2026.1.4
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
dlib @ file:///C:/bld/dlib-split_1761791682434/work
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.1
face-recognition==1.3.0
face_recognition_models==0.3.0
fastapi==0.128.0
fastapi-mail==1.6.1
ffmpy==1.0.0
filelock==3.20.3
fsspec==2026.1.0
gradio==6.3.0
gradio_client==2.0.3
greenlet==3.3.0
groovy==0.1.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.1
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
numpy @ file:///D:/bld/numpy_1747544591878/work/dist/numpy-2.2.6-cp310-cp310-win_amd64.whl#sha256=af66576ce916c795ffa81be9620c8138a3fda6e1713b2bf7775a630e47eca3a8
opencv-python==4.12.0.88
orjson==3.11.5
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pillow==12.1.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.21
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
rich==14.2.0
rsa==4.9.1
safehttpx==0.1.7
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
SQLAlchemy==2.0.45
starlette==0.50.0
tomlkit==0.13.3
tqdm==4.67.1
typer==0.21.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uvicorn==0.40.0
